feat: proxy requests to local frontend if needed
Now that we can proxy to multiple backends, we don't need to run any of
the downstream services as the proxy can handle them. However, we still
need to run another proxy to handle the frontend.

Instead, if we fail to find a response in the specification, we can just
send a request to the local frontend to serve it for us. This handles
things like the JavaScript and CSS to actually load the page and means
we don't need anything other than our frontend and `har-proxy`.